week 3 assignment

+++ routes/campsiteRouter.js
@@ -1,12 +1,15 @@
 const express = require('express');
 const Campsite = require('../models/campsite');
 const authenticate = require('../authenticate');
+const verifyAdmin = require('../authenticate');
 
 const campsiteRouter = express.Router();
 
 campsiteRouter.route('/')
+
 .get((req, res, next) => {
     Campsite.find()
+    .populate('comments.author')
     .then(campsites => {
         res.statusCode = 200;
         res.setHeader('Content-Type', 'application/json');
@@ -15,8 +18,9 @@ campsiteRouter.route('/')
     .catch(err => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
+    verifyAdmin.verifyAdmin(req, res, next)
+    }, (req, res, next) => {
     Campsite.create(req.body)
-    .populate('comments.author')
     .then(campsite => {
         console.log('Campsite Created ', campsite);
         res.statusCode = 200;
@@ -30,6 +34,8 @@ campsiteRouter.route('/')
     res.end('PUT operation not supported on /campsites');
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    verifyAdmin.verifyAdmin(req, res, next)
+    }, (req, res, next) => {
     Campsite.deleteMany()
     .then(response => {
         res.statusCode = 200;
@@ -37,7 +43,8 @@ campsiteRouter.route('/')
         res.json(response);
     })
     .catch(err => next(err));
-});
+    }
+)
 
 campsiteRouter.route('/:campsiteId')
 .get((req, res, next) => {
@@ -55,6 +62,8 @@ campsiteRouter.route('/:campsiteId')
     res.end(`POST operation not supported on /campsites/${req.params.campsiteId}`);
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+    verifyAdmin.verifyAdmin(req, res, next)
+    }, (req, res, next) => {
     Campsite.findByIdAndUpdate(req.params.campsiteId, {
         $set: req.body
     }, { new: true })
@@ -66,6 +75,8 @@ campsiteRouter.route('/:campsiteId')
     .catch(err => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    verifyAdmin.verifyAdmin(req, res, next)
+    }, (req, res, next) => {
     Campsite.findByIdAndDelete(req.params.campsiteId)
     .then(response => {
         res.statusCode = 200;
@@ -118,6 +129,8 @@ campsiteRouter.route('/:campsiteId/comments')
     res.end(`PUT operation not supported on /campsites/${req.params.campsiteId}/comments`);
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    verifyAdmin.verifyAdmin(req, res, next)
+}, (req, res, next) => {
     Campsite.findById(req.params.campsiteId)
     .then(campsite => {
         if (campsite) {
@@ -168,52 +181,64 @@ campsiteRouter.route('/:campsiteId/comments/:commentId')
 .put(authenticate.verifyUser, (req, res, next) => {
     Campsite.findById(req.params.campsiteId)
     .then(campsite => {
-        if (campsite && campsite.comments.id(req.params.commentId)) {
-            if (req.body.rating) {
-                campsite.comments.id(req.params.commentId).rating = req.body.rating;
+        if (req.user._id.equals(campsite.comments.id(req.params.commentId).author._id)) {
+            if (campsite && campsite.comments.id(req.params.commentId)) {
+                if (req.body.rating) {
+                    campsite.comments.id(req.params.commentId).rating = req.body.rating;
+                }
+                if (req.body.text) {
+                    campsite.comments.id(req.params.commentId).text = req.body.text;
+                }
+                campsite.save()
+                .then(campsite => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(campsite);
+                })
+                .catch(err => next(err));
+            } else if (!campsite) {
+                err = new Error(`Campsite ${req.params.campsiteId} not found`);
+                err.status = 404;
+                return next(err);
+            } else {
+                err = new Error(`Comment ${req.params.commentId} not found`);
+                err.status = 404;
+                return next(err);
             }
-            if (req.body.text) {
-                campsite.comments.id(req.params.commentId).text = req.body.text;
-            }
-            campsite.save()
-            .then(campsite => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(campsite);
-            })
-            .catch(err => next(err));
-        } else if (!campsite) {
-            err = new Error(`Campsite ${req.params.campsiteId} not found`);
-            err.status = 404;
-            return next(err);
+
         } else {
-            err = new Error(`Comment ${req.params.commentId} not found`);
-            err.status = 404;
-            return next(err);
+            res.statusCode = 403;
+            res.send("Unauthorized Author")
         }
     })
     .catch(err => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    console.log(req)
     Campsite.findById(req.params.campsiteId)
     .then(campsite => {
-        if (campsite && campsite.comments.id(req.params.commentId)) {
-            campsite.comments.id(req.params.commentId).remove();
-            campsite.save()
-            .then(campsite => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(campsite);
-            })
-            .catch(err => next(err));
-        } else if (!campsite) {
-            err = new Error(`Campsite ${req.params.campsiteId} not found`);
-            err.status = 404;
-            return next(err);
+        if (req.user._id.equals(campsite.comments.id(req.params.commentId).author._id)) {
+            if (campsite && campsite.comments.id(req.params.commentId)) {
+                campsite.comments.id(req.params.commentId).remove();
+                campsite.save()
+                .then(campsite => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(campsite);
+                })
+                .catch(err => next(err));
+            } else if (!campsite) {
+                err = new Error(`Campsite ${req.params.campsiteId} not found`);
+                err.status = 404;
+                return next(err);
+            } else {
+                err = new Error(`Comment ${req.params.commentId} not found`);
+                err.status = 404;
+                return next(err);
+            }
         } else {
-            err = new Error(`Comment ${req.params.commentId} not found`);
-            err.status = 404;
-            return next(err);
+            res.statusCode = 403;
+            res.send("Unauthorized Author")
         }
     })
     .catch(err => next(err));
