+++ routes/promotionRouter.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const promotionRouter = express.Router();
 const authenticate = require('../authenticate');
+const verifyAdmin = require('../authenticate');
 
 const Promotion = require('../models/promotion');
 
@@ -15,6 +16,8 @@ promotionRouter.route('/')
     .catch(err => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
+    verifyAdmin.verifyAdmin(req, res, next)
+    }, (req, res, next) => {
     Promotion.create(req.body)
     .then(promotion => {
         console.log('promotion created');
@@ -29,6 +32,8 @@ promotionRouter.route('/')
     res.end('PUT operation not supported on /promotions')
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    verifyAdmin.verifyAdmin(req, res, next)
+    }, (req, res, next) => {
     Promotion.deleteMany()
     .then(response => {
         res.statusCode = 200;
@@ -39,6 +44,7 @@ promotionRouter.route('/')
 })
 
 promotionRouter.route('/:promotionId')
+
 .get((req, res, next) => {
     Promotion.findById(req.params.promotionId)
     .then(promotion => {
@@ -53,6 +59,8 @@ promotionRouter.route('/:promotionId')
     res.end(`POST operation not supported on /promotions/${req.params.promotionId}`);
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+    verifyAdmin.verifyAdmin(req, res, next)
+    }, (req, res, next) => {
     Promotion.findByIdAndUpdate(req.params.promotionId, {
         $set: req.body
     }, { new: true })
@@ -64,6 +72,8 @@ promotionRouter.route('/:promotionId')
     .catch(err => next(err))
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    verifyAdmin.verifyAdmin(req, res, next)
+    }, (req, res, next) => {
     Promotion.findByIdAndDelete(req.params.promotionId)
     .then(response => {
         res.statusCode = 200;
